+++ src/main/java/com/pioneer/web/system/mapper/SysPostMapper.java
@@ -44,7 +44,7 @@ public interface SysPostMapper extends BaseMapper<SysPost> {
      * @param userId 用户ID
      * @return 选中岗位ID列表
      */
-    List<Integer> selectPostListByUserId(Long userId);
+    List<Long> selectPostListByUserId(Long userId);
 
     /**
      * 查询用户所属岗位组
